Fix bug

+++ src/main/java/com/daylight/civilization_fire/common/content/entity/bot/GuardianBot.java
@@ -164,7 +164,7 @@ public final class GuardianBot extends Bot {
     }
 
     @Override
-    public long getMaxEnergy() {
+    public int getMaxEnergy() {
         return 10000;
     }
 }
